+++ src/storm/common/decorators/controller.py
@@ -1,0 +1,10 @@
+def Controller(route):
+    """
+    Mark a class as a controller and associate it with a route.
+    :param route: Base route for the controller.
+    """
+    def wrapper(cls):
+        cls.__route__ = route
+        cls.__is_controller__ = True
+        return cls
+    return wrapper
