+++ src/storm/common/exceptions/exception.py
