+++ src/storm/common/interceptors/interceptor.py
