+++ src/storm/core/middleware/__init__.py
@@ -1,0 +1,1 @@
+from .middleware import Middleware
