+++ tests/test_core.py
@@ -1,10 +1,16 @@
 import pytest
-from storm import Container, Injectable, Controller, Module
+
+from storm.common.decorators.injectable import Injectable
+from storm.common.decorators import Controller
+from storm.core.container import Container
+from storm.core.module import Module
+
 
 @Injectable(singleton=True)
 class ExampleService:
     def get_data(self):
         return "This is some data from the service."
+
 
 @Controller("/example")
 class ExampleController:
@@ -14,12 +20,14 @@ class ExampleController:
     def get(self):
         return self.service.get_data()
 
+
 class ExampleModule(Module):
     def __init__(self):
         super().__init__(
             controllers=[ExampleController],
             providers=[ExampleService],
         )
+
 
 def test_dependency_injection():
     container = Container()
